+++ packages/react-article-components/src/components/body/audio/control-button.js
@@ -4,20 +4,20 @@ import PauseIcon from '../../../assets/body/audio/pause.svg'
 import PlayIcon from '../../../assets/body/audio/play.svg'
 import React from 'react'
 import styled, { css } from 'styled-components'
-import themeConst from '../../../constants/theme'
 import {
   colorGrayscale,
   colorSupportive,
   COLOR_PINK_ARTICLE,
 } from '@twreporter/core/lib/constants/color'
+import { ARTICLE_THEME } from '@twreporter/core/lib/constants/theme'
 
 function getIconColor(themeName) {
   switch (themeName) {
-    case themeConst.article.v2.photo:
+    case ARTICLE_THEME.v2.photo:
       return colorSupportive.main
-    case themeConst.article.v2.pink:
+    case ARTICLE_THEME.v2.pink:
       return COLOR_PINK_ARTICLE.darkPink
-    case themeConst.article.v2.default:
+    case ARTICLE_THEME.v2.default:
     default:
       return colorSupportive.heavy
   }
@@ -25,17 +25,17 @@ function getIconColor(themeName) {
 
 function getIconWrapperStyles(themeName) {
   switch (themeName) {
-    case themeConst.article.v2.photo:
+    case ARTICLE_THEME.v2.photo:
       return css`
         background-color: ${COLOR_PINK_ARTICLE.darkBlue};
         border-color: ${colorGrayscale.gray400};
       `
-    case themeConst.article.v2.pink:
+    case ARTICLE_THEME.v2.pink:
       return css`
         background-color: ${colorGrayscale.gray100};
         border-color: ${colorGrayscale.gray300};
       `
-    case themeConst.article.v2.default:
+    case ARTICLE_THEME.v2.default:
     default:
       return css`
         background-color: ${colorGrayscale.gray100};
@@ -45,7 +45,7 @@ function getIconWrapperStyles(themeName) {
 }
 
 const IconWrapper = styled.button`
-  ${props => getIconWrapperStyles(props.theme.name)}
+  ${(props) => getIconWrapperStyles(props.theme.name)}
   ${mq.tabletAndBelow`
     width: 48px;
     height: 48px;
@@ -67,9 +67,9 @@ const IconWrapper = styled.button`
     top: 50%;
     transform: translate(-50%, -50%);
     path {
-      stroke: ${props =>
+      stroke: ${(props) =>
         props.$isPlaying ? getIconColor(props.theme.name) : 'none'};
-      fill: ${props =>
+      fill: ${(props) =>
         props.$isPlaying ? 'none' : getIconColor(props.theme.name)};
     }
   }